+++ tests/conftest.py
@@ -7,6 +7,7 @@ from typing import Callable, Optional, List, Any
 
 import pytest
 from click.testing import CliRunner, Result
+from reduct import Bucket, EntryInfo, Record, BucketSettings
 
 from reduct_cli.cli import cli
 
@@ -41,3 +42,52 @@ def _make_url() -> str:
 @pytest.fixture(name="set_alias")
 def _set_alias(runner, conf, url):
     runner(f"-c {conf} alias add test", input=f"{url}\ntoken\n")
+
+
+@pytest.fixture(name="src_settings")
+def _make_settings() -> BucketSettings:
+    return BucketSettings()
+
+
+@pytest.fixture(name="records")
+def _make_records() -> List[Record]:
+    def make_record(timestamp: int, data: bytes) -> Record:
+        async def read_all():
+            return data
+
+        async def read(_n: int):
+            yield data
+
+        return Record(
+            timestamp=timestamp, size=len(data), last=True, read_all=read_all, read=read
+        )
+
+    return [make_record(1000000000, b"Hey"), make_record(5000000000, b"Bye")]
+
+
+@pytest.fixture(name="src_bucket")
+def _make_src_bucket(mocker, src_settings, records) -> Bucket:
+    bucket = mocker.Mock(spec=Bucket)
+    bucket.name = "src_bucket"
+    bucket.get_settings.return_value = src_settings
+    bucket.get_entry_list.return_value = [
+        EntryInfo(
+            name="entry-1",
+            size=1050000,
+            block_count=1,
+            record_count=2,
+            oldest_record=1000000000,
+            latest_record=5000000000,
+        )
+    ]
+
+    bucket.query.return_value = AsyncIter(records)
+    return bucket
+
+
+@pytest.fixture(name="dest_bucket")
+def _make_dest_bucket(mocker) -> Bucket:
+    bucket = mocker.Mock(spec=Bucket)
+    bucket.name = "dest_bucket"
+
+    return bucket